Closes #60
Move remaining requests into requests classes
