+++ backend/src/contracts/Requests/TagRequest.cs
@@ -1,0 +1,4 @@
+namespace contracts.Requests;
+
+public record CreateTagRequest(string Name);
+public record UpdateTagRequest(string Name);
